Add files via upload
A demand forecast system using machine learning. Using TensorFlow to predict taxicab demand on a day-by-day basis. In order to do that, we will first need a training dataset that consists of the factors that affect taxi demand ("input features") and the actual demand ("target") on days in the past. We will use this data to train a neural network that can then predict the demand on days in the future given those features as input.

In this lab, you will use BigQuery in Datalab to build the dataset that we will use in the next lab to carry out the machine learning.